# 올리기 테스트

+++ src/main/java/com/example/movie/service/BoardService.java
@@ -5,11 +5,16 @@ import com.example.movie.entity.Board;
 import com.example.movie.repository.BoardRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.Query;
+import jakarta.persistence.TypedQuery;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.IntStream;
 
 @Service
 public class BoardService {
@@ -58,9 +63,20 @@ public class BoardService {
         boardRepository.save(board);
     }
 
-    public void delete(Long id) {
-        boardRepository.deleteById(id);
-    }
+
+//    public void delete(Long id) {
+//        boardRepository.deleteById(id);
+//    }
+
+        public void delete(Long id) {
+            Board board = boardRepository.findById(id).orElse(null);
+            if (board != null) {
+                board.setMovies(null); // movies 필드를 null로 설정
+                board.setUser(null);   // user 필드를 null로 설정
+                boardRepository.deleteById(id); // board 엔티티 삭제
+            }
+        }
+
 
 //    public List<BoardDto> findAll() {
 //        List<BoardDto> boardDtoList = new ArrayList<>();
@@ -77,5 +93,33 @@ public class BoardService {
         Query query = em.createQuery(sql);
         List<Board> boardList = query.getResultList();
         return boardList;
+    }
+
+    public Page<Board> pageList(Pageable pageable) {
+        return boardRepository.findAll(pageable);
+
+    }
+
+    private static final int BAR_LENGTH=5;
+    public List<Integer> getPaginationBarNumbers(int pageNumber, int totalPage) {
+        int startNumber = Math.max(pageNumber-(BAR_LENGTH/2),0);
+
+        int endNumber = Math.min(startNumber + BAR_LENGTH, totalPage);
+
+        return IntStream.range(startNumber,endNumber).boxed().toList();
+    }
+
+    public BoardDto getOneBoard(Long boardId) {
+        return boardRepository.findById(boardId)
+                .map(x->BoardDto.fromBoardEntity(x))
+                .orElse(null);
+    }
+
+    public List<Board> movie() {
+        String sql = "SELECT b.movieTitle  FROM Board b";
+        TypedQuery<Board> query = em.createQuery(sql, Board.class);
+
+        List<Board> movie = query.getResultList();
+        return movie;
     }
 }
